Implemented Worgen & Goblin custom races
The races are now playable including:
customization
multiple classes
no issue with languages

Some skills needed for weapons/armor are still missing.

+++ src/server/game/Globals/ObjectMgr.h
@@ -659,8 +659,8 @@ typedef std::vector<PlayerCreateInfoItem> PlayerCreateInfoItems;
 
 struct PlayerClassLevelInfo
 {
-    uint16 basehealth = 0;
-    uint16 basemana = 0;
+    uint32 basehealth = 0;
+    uint32 basemana = 0;
 };
 
 struct PlayerClassInfo
